+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.5.0] - 2026-01-04
+
+### Added
+
+- Encrypted data file with passphrase prompts on each run.
+- `TTT_PASSPHRASE` support for non-interactive usage.
+
 ## [v0.4.0] - 2026-01-04
 
 ### Added